vmware-solution/logan_collectors/requirements.txt
oci==2.144.0
pyvmomi==8.0.3.0.1
PyYAML>=6.0
requests>=2.27.0
urllib3>=1.26
python-dateutil>=2.8.2
pytz>=2024.1
tenacity>=8.2.2
